Add `Ssz.hash_tree_root!` function
Unify Beacon.Helper and StateTransition.Misc modules
