+++ lib/lambda_ethereum_consensus/fork_choice/supervisor.ex
@@ -53,16 +53,15 @@ defmodule LambdaEthereumConsensus.ForkChoice do
   end
 
   defp get_latest_block_hash(anchor_state) do
-    with {:ok, state_root} <- Ssz.hash_tree_root(anchor_state) do
-      # The latest_block_header.state_root was zeroed out to avoid circular dependencies
-      anchor_state.latest_block_header
-      |> Map.put(:state_root, state_root)
-      |> Ssz.hash_tree_root()
-    end
+    state_root = Ssz.hash_tree_root!(anchor_state)
+    # The latest_block_header.state_root was zeroed out to avoid circular dependencies
+    anchor_state.latest_block_header
+    |> Map.put(:state_root, state_root)
+    |> Ssz.hash_tree_root!()
   end
 
   defp fetch_anchor_block(%SszTypes.BeaconState{} = anchor_state) do
-    {:ok, block_root} = get_latest_block_hash(anchor_state)
+    block_root = get_latest_block_hash(anchor_state)
 
     case BlockStore.get_block(block_root) do
       {:ok, anchor_block} ->
